sonic-yang-mgmt: uses clause with leaf-list not honored (PR #21907)
When a uses clause imports a grouping, it was only processing leaf
entries and ignoring leaf-list.  That means that for instance in
bgp route maps, route_map_in and route_map_out validations would
fail.

This fixes that behavior and adds test cases to ensure it doesn't
regress in the future.

Signed-off-by: Brad House (@bradh352)