sonic-yang-mgmt: uses clause with leaf-list, choice not honored
When a uses clause imports a grouping, it was only processing leaf
entries and ignoring leaf-list and choice clauses.  That means that
for instance in bgp route maps, route_map_in and route_map_out
validations would fail.

Honoring the uses `refine` clause is now also honored which is
depended upon in sonic-utilities.

This now precompiles the uses clause and integrates it as if the
uses clause was not part of the schema as multiple end users were
having to do this additional processing.

This fixes that behavior and adds test cases to ensure it doesn't
regress in the future.

Signed-off-by: Brad House (@bradh352)